requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.1
mistralai==0.1.3
better-profanity==0.7.0
pydantic==2.6.1
sqlalchemy==2.0.25
mysql-connector-python
